requirements.txt
clearbit==0.1.7
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
